增加地址报错功能，并且解决了报错行数不正确的bug

+++ qtproj/hexDecode.h
@@ -39,5 +39,5 @@ public:
     void Clear(void);
 
 private:
-    uint8_t toUInt(QByteArray str);
+    uint32_t toUInt(QByteArray str);
 };
